PODS-7534 Combine controllers
