+++ app/controllers/register/establishers/company/director/TrusteesAlsoDirectorsController.scala
@@ -34,12 +34,13 @@ package controllers.register.establishers.company.director
 
 import controllers.Retrievals
 import controllers.actions._
-import forms.dataPrefill.DataPrefillCheckboxFormProvider
+import forms.dataPrefill.{DataPrefillCheckboxFormProvider, DataPrefillRadioFormProvider}
 import identifiers.SchemeNameId
 import identifiers.register.establishers.company.CompanyDetailsId
-import identifiers.register.establishers.company.director.TrusteesAlsoDirectorsId
+import identifiers.register.establishers.company.director.{TrusteeAlsoDirectorId, TrusteesAlsoDirectorsId}
+import models._
+import models.prefill.IndividualDetails
 import models.requests.DataRequest
-import models.{CompanyDetails, DataPrefillCheckbox, Index, NormalMode}
 import navigators.Navigator
 import play.api.data.Form
 import play.api.i18n.{I18nSupport, Messages, MessagesApi}
@@ -60,7 +61,8 @@ class TrusteesAlsoDirectorsController @Inject()(override val messagesApi: Messag
                                                 getData: DataRetrievalAction,
                                                 allowAccess: AllowAccessActionProvider,
                                                 requireData: DataRequiredAction,
-                                                formProvider: DataPrefillCheckboxFormProvider,
+                                                formProviderCheckBox: DataPrefillCheckboxFormProvider,
+                                                formProviderRadio: DataPrefillRadioFormProvider,
                                                 dataPrefillService: DataPrefillService,
                                                 val controllerComponents: MessagesControllerComponents,
                                                 val checkBoxView: dataPrefillCheckbox,
@@ -68,56 +70,102 @@ class TrusteesAlsoDirectorsController @Inject()(override val messagesApi: Messag
                                                )(implicit val executionContext: ExecutionContext) extends FrontendBaseController
   with I18nSupport with Retrievals with Enumerable.Implicits {
 
-  private def renderCheckBoxView(status: Status,
-                         ua: UserAnswers,
-                         form: Form[List[Int]],
-                         establisherIndex:Int,
+
+  private def renderView(status: Status,
+                         seqTrustee: Seq[IndividualDetails],
+                         formCheckBox: => Form[List[Int]],
+                         formRadio: => Form[Int],
+                         establisherIndex: Int,
                          companyName: CompanyDetails,
                          schemeName: String)(implicit request: DataRequest[AnyContent]): Future[Result] = {
-    val seqTrustee = dataPrefillService.getListOfTrusteesToBeCopied(establisherIndex)(ua)
-    val pageHeading = Messages("messages__directors__prefill__title")
-    val titleMessage = Messages("messages__directors__prefill__heading", companyName.companyName)
-    val options = DataPrefillCheckbox.checkboxes(seqTrustee)
-    val postCall = controllers.register.establishers.company.director.routes.TrusteesAlsoDirectorsController.onSubmit(establisherIndex)
-    Future.successful(status(checkBoxView(form, Some(schemeName), pageHeading, titleMessage, options, postCall)))
+    if (seqTrustee.size > 1) {
+      val pageHeading = Messages("messages__directors__prefill__title")
+      val titleMessage = Messages("messages__directors__prefill__heading", companyName.companyName)
+      val options = DataPrefillCheckbox.checkboxes(seqTrustee)
+      val postCall = controllers.register.establishers.company.director.routes.TrusteesAlsoDirectorsController.onSubmit(establisherIndex)
+      Future.successful(status(checkBoxView(formCheckBox, Some(schemeName), pageHeading, titleMessage, options, postCall)))
+    } else {
+      val pageHeading = Messages("messages__directors__prefill__title")
+      val titleMessage = Messages("messages__directors__prefill__heading", companyName.companyName)
+      val options = DataPrefillRadio.radios(seqTrustee)
+      val postCall = controllers.register.establishers.company.director.routes.TrusteesAlsoDirectorsController.onSubmit(establisherIndex)
+      Future.successful(status(radioView(formRadio, Some(schemeName), pageHeading, titleMessage, options, postCall)))
+    }
   }
 
   def onPageLoad(establisherIndex: Index): Action[AnyContent] =
     (authenticate() andThen getData(NormalMode, None) andThen allowAccess(None) andThen requireData).async {
       implicit request =>
         (CompanyDetailsId(establisherIndex) and SchemeNameId).retrieve.right.map { case companyName ~ schemeName =>
-          renderCheckBoxView(Ok, request.userAnswers,
-            formCheckBox(establisherIndex)(request.userAnswers, implicitly), establisherIndex, companyName, schemeName)
+          val seqTrustee: Seq[IndividualDetails] = dataPrefillService.getListOfTrusteesToBeCopied(establisherIndex)(request.userAnswers)
+          renderView(Ok,
+            seqTrustee,
+            formCheckBox(establisherIndex)(request.userAnswers, implicitly),
+            formRadio(establisherIndex)(request.userAnswers, implicitly),
+            establisherIndex,
+            companyName,
+            schemeName
+          )
         }
     }
 
+  //scalastyle:off method.length
   def onSubmit(establisherIndex: Index): Action[AnyContent] =
     (authenticate() andThen getData(NormalMode, None) andThen allowAccess(None) andThen requireData).async {
       implicit request =>
-        implicit val ua: UserAnswers = request.userAnswers
+        val seqTrustee: Seq[IndividualDetails] = dataPrefillService.getListOfTrusteesToBeCopied(establisherIndex)(request.userAnswers)
         (CompanyDetailsId(establisherIndex) and SchemeNameId).retrieve.right.map { case companyName ~ schemeName =>
-          formCheckBox(establisherIndex).bindFromRequest().fold(
-            (formWithErrors: Form[List[Int]]) => {
-              renderCheckBoxView(BadRequest, request.userAnswers, formWithErrors, establisherIndex, companyName, schemeName)
-            },
-            value => {
-              def uaAfterCopy: UserAnswers = (if (value.headOption.getOrElse(-1) < 0) {
-                ua
-              } else {
-                dataPrefillService.copyAllTrusteesToDirectors(ua, value, establisherIndex)
-              }).setOrException(TrusteesAlsoDirectorsId(establisherIndex))(value)
-
-              userAnswersService.upsert(NormalMode, None, uaAfterCopy.json).map { _ =>
-                Redirect(navigator.nextPage(TrusteesAlsoDirectorsId(establisherIndex), NormalMode, uaAfterCopy, None))
-              }
+          if (seqTrustee.size > 1) {
+            val boundForm: Form[List[Int]] = formCheckBox(establisherIndex)(request.userAnswers, implicitly).bindFromRequest()
+            boundForm.value match {
+              case Some(value) if boundForm.errors.isEmpty =>
+                def uaAfterCopy: UserAnswers = (if (value.headOption.getOrElse(-1) < 0) {
+                  request.userAnswers
+                } else {
+                  dataPrefillService.copyAllTrusteesToDirectors(request.userAnswers, value, establisherIndex)
+                }).setOrException(TrusteesAlsoDirectorsId(establisherIndex))(value)
+                userAnswersService.upsert(NormalMode, None, uaAfterCopy.json).map { _ =>
+                  Redirect(navigator.nextPage(TrusteesAlsoDirectorsId(establisherIndex), NormalMode, uaAfterCopy, None))
+                }
+              case _ =>
+                renderView(BadRequest,
+                  seqTrustee,
+                  boundForm,
+                  formRadio(establisherIndex)(request.userAnswers, implicitly),
+                  establisherIndex,
+                  companyName,
+                  schemeName
+                )
             }
-          )
+          } else {
+            val boundForm = formRadio(establisherIndex)(request.userAnswers, implicitly).bindFromRequest()
+            boundForm.value match {
+              case Some(value) if boundForm.errors.isEmpty =>
+                def uaAfterCopy: UserAnswers = (if (value < 0) {
+                  request.userAnswers
+                } else {
+                  dataPrefillService.copyAllTrusteesToDirectors(request.userAnswers, Seq(value), establisherIndex)
+                }).setOrException(TrusteeAlsoDirectorId(establisherIndex))(value)
+                userAnswersService.upsert(NormalMode, None, uaAfterCopy.json).map { _ =>
+                  Redirect(navigator.nextPage(TrusteeAlsoDirectorId(establisherIndex), NormalMode, uaAfterCopy, None))
+                }
+              case _ =>
+                renderView(BadRequest,
+                  seqTrustee,
+                  formCheckBox(establisherIndex)(request.userAnswers, implicitly),
+                  boundForm,
+                  establisherIndex,
+                  companyName,
+                  schemeName
+                )
+            }
+          }
         }
     }
 
   private def formCheckBox(index: Index)(implicit ua: UserAnswers, messages: Messages): Form[List[Int]] = {
     val existingDirCount = ua.allDirectorsAfterDelete(index).size
-    formProvider(
+    formProviderCheckBox(
       existingDirCount,
       "messages__directors__prefill__multi__error__required",
       "messages__directors__prefill__multi__error__noneWithValue",
@@ -125,4 +173,7 @@ class TrusteesAlsoDirectorsController @Inject()(override val messagesApi: Messag
         existingDirCount, /*config.maxDirectors*/ 10 - existingDirCount)
     )
   }
+
+  private def formRadio(index: Index)(implicit ua: UserAnswers, messages: Messages): Form[Int] =
+    formProviderRadio("messages__directors__prefill__single__error__required")
 }
