+++ conf/establishersCompanyDirector.routes
@@ -134,8 +134,5 @@ POST        /:establisherIndex/:mode/company/director/:directorIndex/confirm-del
 GET         /:establisherIndex/company/director/:directorIndex/already-deleted                                           controllers.register.establishers.company.director.AlreadyDeletedController.onPageLoad(establisherIndex: Index, directorIndex: Index, srn: Option[String] = None)
 GET         /:establisherIndex/company/director/:directorIndex/already-deleted/:srn                                      controllers.register.establishers.company.director.AlreadyDeletedController.onPageLoad(establisherIndex: Index, directorIndex: Index, srn: Option[String])
 
-GET         /:establisherIndex/company/director/trustee-also-director                         controllers.register.establishers.company.director.TrusteeAlsoDirectorController.onPageLoad(establisherIndex: Index)
-POST        /:establisherIndex/company/director/trustee-also-director                          controllers.register.establishers.company.director.TrusteeAlsoDirectorController.onSubmit(establisherIndex: Index)
-
 GET         /:establisherIndex/company/director/trustees-also-directors                         controllers.register.establishers.company.director.TrusteesAlsoDirectorsController.onPageLoad(establisherIndex: Index)
 POST        /:establisherIndex/company/director/trustees-also-directors                          controllers.register.establishers.company.director.TrusteesAlsoDirectorsController.onSubmit(establisherIndex: Index)
